Refactored middleware functionality
